+++ index.js
@@ -46,7 +46,7 @@ const enrichTestData = (data, base_data_path) =>
 /**
  * Make a test case(s) using a collection of test data objects and
  * an unary function receiving every item of the collection possible
- * updated w/ data loaded from JSON files.
+ * updated w/ data loaded from JSON file(s).
  *
  * The only mandatory property for the test item is the `name` --
  * a short test case description.
@@ -76,4 +76,33 @@ const makeTestCase = (tests, fn, is_async=false) =>
     });
 }
 
+/**
+ * Make a test suite.
+ *
+ * Same as `makeTestCase`, but use `describe()` instead of `it()` from `mocha`.
+ */
+const makeTestSuite = (tests, fn, is_async=false) =>
+{
+    tests.forEach(test =>
+    {
+        if (is_async)
+        {
+            describe(test.name, async function()
+            {
+                // Ok, call a given function passing collected test data
+                await fn(enrichTestData(test, this.parent.file));
+            });
+        }
+        else
+        {
+            describe(test.name, function()
+            {
+                // Ok, call a given function passing collected test data
+                fn(enrichTestData(test, this.parent.file));
+            });
+        }
+    });
+}
+
 module.exports = makeTestCase;
+module.exports.makeTestSuite = makeTestSuite;